+++ platform/android/java/lib/src/org/godotengine/godot/input/GodotTextInputWrapper.java
@@ -94,17 +94,15 @@ public class GodotTextInputWrapper implements TextWatcher, OnEditorActionListene
 	public void beforeTextChanged(final CharSequence pCharSequence, final int start, final int count, final int after) {
 		//Log.d(TAG, "beforeTextChanged(" + pCharSequence + ")start: " + start + ",count: " + count + ",after: " + after);
 
-		mRenderView.queueOnRenderThread(() -> {
-			for (int i = 0; i < count; ++i) {
-				GodotLib.key(KeyEvent.KEYCODE_DEL, KeyEvent.KEYCODE_DEL, 0, true);
-				GodotLib.key(KeyEvent.KEYCODE_DEL, KeyEvent.KEYCODE_DEL, 0, false);
+		for (int i = 0; i < count; ++i) {
+			GodotLib.key(KeyEvent.KEYCODE_DEL, KeyEvent.KEYCODE_DEL, 0, true);
+			GodotLib.key(KeyEvent.KEYCODE_DEL, KeyEvent.KEYCODE_DEL, 0, false);
 
-				if (mHasSelection) {
-					mHasSelection = false;
-					break;
-				}
+			if (mHasSelection) {
+				mHasSelection = false;
+				break;
 			}
-		});
+		}
 	}
 
 	@Override
@@ -115,17 +113,15 @@ public class GodotTextInputWrapper implements TextWatcher, OnEditorActionListene
 		for (int i = start; i < start + count; ++i) {
 			newChars[i - start] = pCharSequence.charAt(i);
 		}
-		mRenderView.queueOnRenderThread(() -> {
-			for (int i = 0; i < count; ++i) {
-				int key = newChars[i];
-				if ((key == '\n') && !mEdit.isMultiline()) {
-					// Return keys are handled through action events
-					continue;
-				}
-				GodotLib.key(0, 0, key, true);
-				GodotLib.key(0, 0, key, false);
+		for (int i = 0; i < count; ++i) {
+			int key = newChars[i];
+			if ((key == '\n') && !mEdit.isMultiline()) {
+				// Return keys are handled through action events
+				continue;
 			}
-		});
+			GodotLib.key(0, 0, key, true);
+			GodotLib.key(0, 0, key, false);
+		}
 	}
 
 	@Override
@@ -133,13 +129,11 @@ public class GodotTextInputWrapper implements TextWatcher, OnEditorActionListene
 		if (mEdit == pTextView && isFullScreenEdit()) {
 			final String characters = pKeyEvent.getCharacters();
 
-			mRenderView.queueOnRenderThread(() -> {
-				for (int i = 0; i < characters.length(); i++) {
-					final int ch = characters.codePointAt(i);
-					GodotLib.key(0, 0, ch, true);
-					GodotLib.key(0, 0, ch, false);
-				}
-			});
+			for (int i = 0; i < characters.length(); i++) {
+				final int ch = characters.codePointAt(i);
+				GodotLib.key(0, 0, ch, true);
+				GodotLib.key(0, 0, ch, false);
+			}
 		}
 
 		if (pActionID == EditorInfo.IME_ACTION_DONE) {
